# Add `consensus_decode_from_finite_reader` optimization