@dpc 2022-05-29
OK. I have to benchmark this on a laptop, with a lot of noise, so I've ran the bench multiple times and picked some best results for both `master` and this PR:

```
old:
test blockdata::block::benches::bench_block_deserialize                 ... bench:   1,381,111 ns/iter (+/- 18,765)
test blockdata::transaction::benches::bench_transaction_deserialize     ... bench:         133 ns/iter (+/- 5)

test blockdata::block::benches::bench_block_deserialize                 ... bench:   1,431,350 ns/iter (+/- 187,241)
test blockdata::transaction::benches::bench_transaction_deserialize     ... bench:         177 ns/iter (+/- 4)

test blockdata::block::benches::bench_block_deserialize                 ... bench:   1,364,857 ns/iter (+/- 22,960)
test blockdata::transaction::benches::bench_transaction_deserialize     ... bench:         127 ns/iter (+/- 6)


new:
test blockdata::block::benches::bench_block_deserialize                 ... bench:   1,112,535 ns/iter (+/- 45,702)
test blockdata::transaction::benches::bench_transaction_deserialize     ... bench:         145 ns/iter (+/- 3)

test blockdata::block::benches::bench_block_deserialize                 ... bench:   1,093,602 ns/iter (+/- 21,832)
test blockdata::transaction::benches::bench_transaction_deserialize     ... bench:          97 ns/iter (+/- 2)

test blockdata::block::benches::bench_block_deserialize                 ... bench:   1,133,440 ns/iter (+/- 36,818)
test blockdata::transaction::benches::bench_transaction_deserialize     ... bench:         146 ns/iter (+/- 2)
```