@tcharding 2022-05-31
I did a bunch of docs fixes (mainly grammar), if you'd like to use them I pushed a patch to branch `dpc-issue-997-2` on my fork.

I had one question, just out of interest, is there a reason you use such short line length in the comments? 

While cleaning up I've been favouring line length 100 for comments (I try to not touch them if they are at least 80 though if changing them does not make them noticeably nicer.)