@dpc 2022-05-31
> I had one question, just out of interest, is there a reason you use such short line length in the comments?

I typically let `rustfmt` deal with it, so I wasn't thinking much about. I'm so used to typing out garbage that `rustfmt` just fixes for me, that I have a hard enough time with the code, and wasn't even thinking about comments. :D

> I did a bunch of docs fixes (mainly grammar), if you'd like to use them I pushed a patch to branch `dpc-issue-997-2` on my fork.

Notes. Will do. Thanks a lot!