@tcharding 2022-05-31
> I typically let rustfmt deal with it, so I wasn't thinking much about. I'm so used to typing out garbage that rustfmt just fixes for me, that I have a hard enough time with the code, and wasn't even thinking about comments. :D

ah, no sweat. We should have that fixed up soon hopefully :)