@apoelstra 2022-05-31
No comment on choice of `assert` ;)