@dpc 2022-06-12
@RCasatta since I know you're using these traits in [`blocks_iterator`](https://github.com/RCasatta/blocks_iterator) and care about performance there you might want to be aware of this and #1035 .